backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
psycopg[binary]==3.2.3
python-dotenv==1.0.0
Werkzeug==3.0.1